requirements.txt
py-cord
audioop-lts
PyNaCl
pymongo
yandex-music
pillow
python-dotenv
wavelink